+++ adminer/drivers/pgsql.inc.php
@@ -135,6 +135,7 @@ if (isset($_GET["pgsql"])) {
 				$return->orgtable = pg_field_table($this->result, $column);
 				$return->name = pg_field_name($this->result, $column);
 				$type = pg_field_type($this->result, $column);
+				$return->native_type = $type;
 				$return->type = (preg_match(number_type(), $type) ? 0 : 15);
 				$return->charsetnr = ($type == "bytea" ? 63 : 0); // 63 - binary
 				return $return;