+++ adminer/include/editing.inc.php
@@ -53,7 +53,7 @@ function print_select_result($result, ?Db $connection2 = null, array $orgtables 
 					$blobs[$j] = true;
 				}
 				$types[$j] = $field->type;
-				echo "<th" . ($orgtable != "" || $field->name != $orgname ? " title='" . h(($orgtable != "" ? "$orgtable." : "") . $orgname) . "'" : "") . ">" . h($name)
+				echo "<th title='" . h(trim(($orgtable != "" ? "$orgtable.$orgname" : ($field->name != $orgname ? $orgname : "")) . " " . driver()->typeName($field))) . "'>" . h($name)
 					. ($orgtables ? doc_link(array(
 						'sql' => "explain-output.html#explain_" . strtolower($name),
 						'mariadb' => "explain/#the-columns-in-explain-select",